Added Temporary Freeze to Items

+++ entities/entities/lia_item/init.lua
@@ -10,10 +10,16 @@ function ENT:Initialize()
     self:SetCollisionGroup(COLLISION_GROUP_WEAPON)
     self.health = 50
     local physObj = self:GetPhysicsObject()
-    if IsValid(physObj) then
-        physObj:EnableMotion(true)
-        physObj:Wake()
-    end
+    ent:EnableMotion(false)
+    timer.Simple(
+        3,
+        function()
+            if IsValid(physObj) then
+                physObj:EnableMotion(true)
+                physObj:Wake()
+            end
+        end
+    )
 
     hook.Run("OnItemSpawned", self)
 end
@@ -36,10 +42,7 @@ function ENT:setItem(itemID)
     if not itemTable then return self:Remove() end
     itemTable:sync()
     local model = itemTable.onGetDropModel and itemTable:onGetDropModel(self) or itemTable:getModel() or itemTable.model
-    if itemTable.worldModel then
-        model = itemTable.worldModel == true and "models/props_junk/cardboard_box004a.mdl" or itemTable.worldModel
-    end
-
+    if itemTable.worldModel then model = itemTable.worldModel == true and "models/props_junk/cardboard_box004a.mdl" or itemTable.worldModel end
     self:SetModel(model)
     self:SetSkin(itemTable.skin or 0)
     self:SetMaterial(itemTable.material or "")
@@ -49,10 +52,7 @@ function ENT:setItem(itemID)
     self:setNetVar("id", itemTable.uniqueID)
     self:setNetVar("instanceID", itemTable.id)
     self.liaItemID = itemID
-    if table.Count(itemTable.data) > 0 then
-        self:setNetVar("data", itemTable.data)
-    end
-
+    if table.Count(itemTable.data) > 0 then self:setNetVar("data", itemTable.data) end
     local physObj = self:GetPhysicsObject()
     if not IsValid(physObj) then
         local min, max = Vector(-8, -8, -8), Vector(8, 8, 8)
@@ -65,9 +65,7 @@ function ENT:setItem(itemID)
         physObj:Wake()
     end
 
-    if itemTable.onEntityCreated then
-        itemTable:onEntityCreated(self)
-    end
+    if itemTable.onEntityCreated then itemTable:onEntityCreated(self) end
 end
 --------------------------------------------------------------------------------------------------------
 function ENT:breakEffects()
@@ -84,23 +82,17 @@ function ENT:OnRemove()
     local itemTable = self:getItemTable()
     if self.breaking then
         self:breakEffects()
-        if itemTable and itemTable.onDestroyed then
-            itemTable:onDestroyed(self)
-        end
-
+        if itemTable and itemTable.onDestroyed then itemTable:onDestroyed(self) end
         self.breaking = false
     end
 
-    if not lia.shuttingDown and not self.liaIsSafe and self.liaItemID then
-        lia.item.deleteByID(self.liaItemID)
-    end
+    if not lia.shuttingDown and not self.liaIsSafe and self.liaItemID then lia.item.deleteByID(self.liaItemID) end
 end
 --------------------------------------------------------------------------------------------------------
 function ENT:Think()
     local itemTable = self:getItemTable()
     if itemTable and itemTable.think then return itemTable:think(self) end
     self:NextThink(CurTime() + 1)
-
     return true
 end
 --------------------------------------------------------------------------------------------------------
